Changed folder Names, and other changes
