+++ HotFix Generator/bl3data.py
@@ -189,7 +189,7 @@ class BL3Data(object):
             #Unfortuanutly this is not public as it needs my credentials to login,
             #but if it becomes popular enough I will try and make it a public database that is seperate from mine so that everyone can use
             self.db = pyodbc.connect('Driver={SQL Server};'
-                                     'Server=A-LARGE-MANS-PC\SQLEXPRESS;'
+                                     'Server=<SQLSERVER>;'
                                      'Database=bl3references;'
                                      'Trusted_Connection=yes;')
             self.curs = self.db.cursor()
