+++ HotFix Generator/notes.txt
@@ -1,0 +1,36 @@
+This section is more so for my purposes than anything else
+Its jsut a nice reminder/convient place to copy stuff if I need to use it
+
+"""
+Mod types
+Mod.PATCH - Will create a SparkPatchEntry hotfix
+Mod.LEVEL- Will create a SparkLevelPatchEntry hotfix
+Mod.EARLYLEVEL- Will create a SparkEarlyLevelPatchEntry hotfix
+Mod.CHAR - Will create a SparkCharacterLoadedEntry hotfix
+#bottom onees are currently not used, do not use them
+Mod.PACKAGE - Will create a SparkStreamedPackageEntry hotfix
+Mod.POST - Will create a SparkPostLoadedEntry hotfix
+"""
+#regular hotfix layout
+"""
+1. What type of patch it is
+2. What map location is this located in (can leave blank or have MatchAll)
+3. Where file directory of the json file you want to grab info of + the json file. EX: /Game/GameData/Regions/RegionManagerData + RegionManagerData
+4. data to manipulate? look more into this later. In the json file, it is the "NAME" :, This is the data you are looking for. may need to make a method to grab all these names, or look to see if there is one
+    4a. sometimes you will need to have an addition comma to put in the function.
+    Look at BL3Redux for idea
+5. Set the value to true
+    5.if you are changing a value of something, set it to an actual number instead of true
+6. leave blank most times
+"""
+
+
+"""
+intake = input("")
+input1 = input("Mod name: ")
+input2 = input("Mod name: ")
+input3 = input("Mod name: ")
+input4 = input("Mod name: ")
+input5 = input("Mod name: ")
+input6 = input("Mod name: ")
+"""
